refactor(reports): keep viewsets.py naming, fold bulk helper back in
This is a sync-DRF → async-ADRF conversion of the existing
ReportViewSet, not a file restructure. Undoes the earlier
extract-into-bulk.py + rename-to-views.py steps:

  - Move bulk_upsert_reports back into radis/reports/api/viewsets.py
    (delete radis/reports/api/bulk.py).
  - Rename radis/reports/api/views.py back to viewsets.py.
  - Point urls.py at `from .viewsets import ReportViewSet`.
  - Update test_bulk_upsert.py and radis-client/tests/test_client.py
    to import / patch under `radis.reports.api.viewsets`.
  - Update spec + plan to match: one module, no file split, no rename.

Behaviour is unchanged from the previous viewset commit (7215e2f9).
CI on that commit was green; the file moves here are import-path
only and should keep CI green.
