+++ radis-client/tests/test_client.py
@@ -19,7 +19,7 @@ def test_report_data_valid():
 def test_report_data_post(live_server: LiveServer, mocker: MockerFixture):
     # Make sure it won't try to save created reports to any full text search database
     # as those are not available during test
-    mocker.patch("radis.reports.api.views.reports_created_handlers", return_value=[])
+    mocker.patch("radis.reports.api.viewsets.reports_created_handlers", return_value=[])
 
     _, _, token = create_admin_with_group_and_token()
     client = RadisClient(live_server.url, token)
