+++ radis/reports/api/urls.py
@@ -1,7 +1,7 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
-from .views import ReportViewSet
+from .viewsets import ReportViewSet
 
 router = DefaultRouter()
 router.register("", ReportViewSet, basename="report")
